Adding support for code coverage.
Adding configuration for Grunt based on the Grunt documentation
https://gruntjs.com/

and the documentation of the code coverage :
https://github.com/jribble/grunt-jasmine-node-coverage

Adding also the code for testing the library. The spec file for the library
is using the jasmine-node syntax. Because jasmine-node supports only jasmine
less than 2.X, there is no support for beforeAll function.